more reorg and cleaning
